Adjust to GenericInstanceFilter API changes
