+++ packages/components/src/presentation-components/tree/DataProvider.ts
@@ -29,7 +29,8 @@ import { Presentation } from "@itwin/presentation-frontend";
 import { createDiagnosticsOptions, DiagnosticsProps } from "../common/Diagnostics";
 import { getRulesetId, RulesetOrId, translate } from "../common/Utils";
 import { PresentationComponentsLoggerCategory } from "../ComponentsLoggerCategory";
-import { PresentationInstanceFilter, PresentationInstanceFilterInfo } from "../instance-filter-builder/PresentationFilterBuilder";
+import { PresentationInstanceFilterInfo } from "../instance-filter-builder/PresentationFilterBuilder";
+import { PresentationInstanceFilter } from "../instance-filter-builder/PresentationInstanceFilter";
 import { IPresentationTreeDataProvider } from "./IPresentationTreeDataProvider";
 import { InfoTreeNodeItemType, isPresentationTreeNodeItem, PresentationTreeNodeItem } from "./PresentationTreeNodeItem";
 import { createInfoNode, createTreeNodeItem, CreateTreeNodeItemProps, pageOptionsUiToPresentation } from "./Utils";
